Separate Checkout interface from transaction code
