+++ packages/dds/tree/src/core/index.ts
@@ -121,7 +121,7 @@ export {
 
 export { ICheckout, TransactionResult } from "../checkout";
 
-export { Checkout, runSynchronousTransaction } from "../transaction";
+export { Checkout } from "../transaction";
 
 export {
     Index,
